Improve "Skip commit/push hook" options
* Invert "Skip" to "Run hook" since positive options are more clear.
* Place the commit option near other code analysis checks.
* Don't remember the state of the checkbox: it is dangerous to skip
  hooks, so it should be a one-time decision each time.

GitSkipHooksCommitHandlerFactory is defined as a CheckinHandlerFactory,
and not VcsCheckinHandlerFactory to be able to place the checkbox
below all other code analysis checkboxes:
since bc02309, VCS handlers are processed before standard handlers,
but that means that their components are placed also before components
of standard handlers.
On the other hand, the visibility of "Run Git hooks" checkbox
is controlled by its handler, so it is OK to make it a standard
CheckinHandler.

Improvement for IDEA-115914.