execution: Restore CommandLineUtil.specialQuote() behavior for non-Windows
Although not used for OS other than Windows in the code base, using that
would emit double quotes for parameter quoted like that. Even though it's
probably not the intended behavior, it's worth restoring it at least to
preserve backward compatibility.

This needs to be reconsidered, since the original purpose of inescapable
quoting is overcoming Windows command line quirks, not passing these quotes
to the target process.