PY-22556, PY-23233, PY-23217: Set working directory for tests that are not in packages
When you want to launch test in folder "tests" which is not a package, you have several troubles:
* Python frameworks report tests relative to this folder (i.e. test_module.foo instead of tests.test_module.foo)
* You can't provide test name as tests.foo.MyTest since "tests" is not package

To solve it, we set "tests" as package. But you can't simply set deepest folder as package: it may break relative imports.
So, newly created configuration should use topmost folder which is not a package.
Consider following layout: tests ( not a package), logic (package), test_module.py
 Here is: "logic.test_module" and "tests" as working directory.

 On Java side, we get working dir reported by python and resolve names against it.
 Several hacks like index lookup of function are not needed now.