Update Git repository info synchronously after fetch
GitRepositoryFiles.refresh() refreshes the VFS, which generate VFS
events, which are handled by GitRepositoryUpdater, which _schedules_
the real update(). So even if the VFS refresh is synchronous, actual
update of the repo info happens later.

It is important to update repository info immediately after fetch,
since this information is used in e.g. GitUpdateProcess.
But we don't want both refresh() and update(), because it will cause
reading .git twice.

Thus introducing a separate method to refresh information which is
not covered by update().

The remaining refresh() method doesn't need the "async" parameter
anymore, since it is async at any case.