Support rejected multi-repo push if the branch in other repo is deleted
Suppose the branch 'feature' is removed in repoA, and got a commit in
repoB. Then push to repoB is rejected. Subsequent update fails,
because the branch is removed in repoA. That shouldn't happen, because
this fact doesn't affect pushing to repoB.

The fix is to provide a flag not to check for such removed branches
during update.

Fixes IDEA-169877.