[git] set git.diff.renameLimit.infinity=false to deal with git index performance problems
On some projects, setting diff.renameLimit to infinity results in git log indexing process to run for several hours and consume a lot of memory. To deal with this problem, we have to turn this option off.
Unfortunately, this results in incomplete history for files that were renamed as a part of a large move (>1k files). We have to fix this problem separately in the future.

diff.renameLimit=0 was first introduced in de41656b7c156c4640e6fffeeeea020deb127c52

IDEA-170443 (indexing), IDEA-123397 (file history through commits with big renames)

(cherry picked from commit f04c9cc)