[git] avoid copying the whole git log output for parsing
Git log output can be quite big, either we have a big list of commits or one big commit with many changed files (the former is not a problem, but the latter can be). In order to parse it GitLogParser is used, but unfortunately its methods required a String as an input, while git log output is stored in a StringBuilder. StringBuilder.toString method copies the contents instead of sharing it. This is unnecessary since parsing is done sequentially. So to avoid input copying GitLogParser now accepts CharSequences.

(cherry picked from commit 7ed158d)