PY-23578 Optimize Imports preserves a blank line before the first import import
It happened due to the bug in the platform and only when the number of
blank lines after the import statement is greater than that before it.

As a temporary workaround before the relevant fix in the platform is
accepted I delete old imports as text through the underlying document.