Update gradle in Intellij to version 3.5
Old version 3.3 of Gradle doesn't contain a new API in Gradle Tooling
API introduced in https://github.com/gradle/gradle/pull/1442.

This API is needed with plugin 2.5 to be able to run Gradle tasks
(assemble artifacts) and fetch a model (outputs of this task) in a
single Gradle invocation, which is necessary for splits and Instant
Apps.

Change-Id: Ia8d00ff50b7d3e4eafbfbcad5d2b3810829eb465