[vcs-log] introduce VcsLogDiffHandler interface and an implementation for git
This class will handle diff from log-based file history.

We could use DiffFromHistoryHandler for this, but the problem is that standart handler can not handle diff for directories (IDEA-170456, IDEA-171150)
and vcs-specific handlers work with vcs-specific instances of VcsFileRevision. There is no api to create such revisions from log,
and it is unclear whether we really want to have it, since VcsFileRevision is such a flawed concept
(this is why we have things like VcsFileRevisionEx). So new, VcsFileRevision-free, handler is introduced
to help with the problem.

(cherry picked from commit d7c135b)