Add studio_presubmit inspection profile to tools/idea
This is supposed to represent a hihgly reduced profile comparing to the
general one used in nightly runs, and should aim to highlight only the
most essential issues in order to reduce PSQ time and not generate noise
in CL comments.

Currently this includes the default subsets of:

+ Java/Probable bugs
+ Java/Imports (amended)
+ Java/Concurrency annotations
+ Java/Memory issues
+ Java/Numeric issues
+ Java/JavaDoc issues
+ Java/Performance issues
+ Java/Threading issues
+ Kotlin
+ Groovy
+ XML
+ Maven
+ Ant
+ UI Form problems
+ RegExp

There might be 1-2 inspections enabled in a couple of other,
more low-profile categories, but there are few of them.

Test: N/A
Change-Id: I1603c3bf7c6b28baa2a00fb3e475705d482b279a