Cython speedups break debugging with Stackless (PY-23283)
The problem appears when compiling Cython extensions with Cython 0.25, and it doesn't appear when compiling them with Cython 0.24. It looks like it's Cython bug, but we can't use old version of Cython, because it isn't compatible with Python 3.6 on Windows. So we have to (temporarily) disable Cython extensions for Stackless despite the fact that its implementation is "CPython".

(cherry picked from commit eaa6f4d)

(cherry picked from commit 32b3fae)