SystemHealthMonitor: fix notification race
Android Studio defines a custom first-run wizard, which exposes
a race condition in displaying notifications in SystemHealthMonitor.

The root problem is that the appFrameCreated callback happens before
the frame is actually visible, see IdeaApplication where we have:

  lifecyclePublisher.appFrameCreated(args, willOpenProject)
  ... followed by, later
  WelcomeFrame.showNow()

And WelcomeFrame.showNow() first goes through extensions before making
itself available:

  if (ourInstance == null) {
    ...
    for (WelcomeFrameProvider provider : EP.getExtensions()) {
      frame = provider.createFrame();
    }
    ...
    ourInstance = frame;
  }

For any non-trivial extension that displays a dialog and starts
consuming events from the AWT queue, this leads to a race condition:
if SystemHealthMonitor has notifications to show, it may not see
any visible frame and throw an NPE because the welcome frame didn't
get a chance to initialize itself.

We propose to use the appStarting callback, which is guaranteed to
happen later in a separate transaction.

Bug: http://b.android.com/241387