Do not parse/serialize additional compiler arguments in JPS
This is mostly a revert of fdc797bc, where the issue of the compiler
failing on repeated command line arguments was worked around. Now the
compiler behaves more properly: it uses the last passed argument and its
value, and prints a warning that several values have been passed. With
this behavior, the workaround is no longer necessary

GitOrigin-RevId: 94b9bcdf5ec2dd5ff0f7730763cfe8a4ec8b04d8