[file-history] new file history algorithm that correctly traverses merge commits where there was a rename in one of the branches
Since now indexes know changed files from all of the parents in merge commits, this information could be utilized to correctly build file history and thus fix some of the issues with it.
Essentially dfs is employed, same as before. But now algorithm works in two passes.

In the first pass, file name in each commit is calculated while doing a dfs on the visible graph.
The trick here is that now for getting changes a commit and a parent index is required (and not just a commit, as before). But in the visible graph, a commit parent is not its real parent in git.
To get a real parent, path between two commits is obtained with bfs.

In the second pass, some commits that have unrelated changes or are trivial merges are thrown out. During this stage some related trivial merges can be thrown out. This is going to be fixed later.

IDEA-170468