[groovy] refactor index access (GrIndexProperty)
The element may appear in rhs as well as in lhs. There is special
case when the same element is both a rValue and a lValue: a[0] += b.
Because of that we need ability to separately resolve getAt() and
putAt() methods. Thus the element cannot be a GrCall.

Added:
- getLValueReference() and getRValueReference();
- getType() always returns type as if element is a rValue.
Removed:
- getGetterType() as it is became unneeded;
- getSetterType() as it was unused already.