Simplify: remove GitCommitWrapper as not needed anymore
When commit is cherry-picked, its commit message can be modified
by user. To hold this new commit message together with the original
commit, the GitCommitWrapper was introduced back in 255403d
to fix IDEA-87375.

However, since 4c6ae37 this logic became unnecessary, because the
standard Commit Dialog is used now, and it passes correct commit
message to CommitExecutors and further by itself.