project model: distinguish libraries imported from external models
This information can be used to show a warning if user tries to manually change roots of a library imported from an external model in the Project Structure dialog (IDEA-171948). This information won't be saved in library configuration files until we implement functionality to store such libraries separately (and enable 'store.imported.project.elements.separately' registry key), because otherwise it would lead to massive changes in files under .idea/libraries.