Remove senseless but brittle test
The test checks that postfix template is NOT expanded by enter if an abbreviation was typed fast and completion lookup was not appeared before pressing Enter.

However, there is nothing wrong in expanding template if lookup was fast enough to appear before pressing Enter.

The second test should be kept, as it checks that Postfix template is expanded no matter is there lookup with template in it or not.