Change log line pattern to include thread identification info.
The existing log line pattern includes both exact date/time of the log
event and the number of milliseconds since the JVM start. This is
basically duplication of the time information.

A more useful entry would be the thread identification information
pointing clearly at the thread which originated the log event. I have
amended this locally when investigating concurrency between
gradle and idea tasks, and it proved to be much more informative. I also
assume this might be useful for any type of concurrency-related
diagnostics conducted by looking at the logs received from the users.

Worth noting that the version of slf4j which idea core is dependent on
(1.7) doesn't yet support the regular thread id specification,
so %t would expand into so-called thread name instead, which
can be quite long and so is forcibly cut to 10 chars here.

In practice, this appears to be enough for contextual identification:

2017-05-18 16:51:09,726 [lease-b736]   INFO - pl.ProjectRootManagerComponent - project roots have changed
2017-05-18 16:51:09,827 [ thread 54]   INFO - penapi.project.DumbServiceImpl - Running dumb mode task: com.android.tools.idea.res.ProjectResourceRepositoryRootListener@13a0da71

Once/if migration to log4j2 happens, thread id should probably still be
used there instead. See the fixed feature request for this:
https://issues.apache.org/jira/browse/LOG4J2-1299

Test: N/A
Change-Id: Idd32eb0c54e014c79fccc0709224e6503f29307e