[tw] tabbed content updates
- Select tab before selecting content in ContentManager

- Do not draw the icon when the tabbed content has only one tab

- Do not show the popup when the tabbed content has only one tab:
Before the change content was selected every time user clicks on
a label. There was no way to simply view the list of available tabs.

- Add reusable list popup steps for selecting content and subcontent.
Before this change dummy DumbAwareAction-s was used to create the list
popup via ActionGroup.
Also this addes icons as a side effect of using custom list popup steps.

- Do not show visible tabs in More action