[file-history] persist commit parents order in the index in order to make it work with git bek parent fixer
Since sometimes git log provider returns commits with reversed parents due to GitBekParentFixer existence, saving just a list of changes into index was insufficient.
Persistent map containing parents list for every commit was added.

GitBekParentFixer is required for the following situation: a user pulls commits from the remote with merge, and then pushes merged branch back. This creates incorrect order of parents for automatic merge commits (changes that were in the remote branch should go first and be displayed on the left). GitBekParentFixer finds such commits, filtering log by message, and reverses parent back for them. This is obviously a hack, but simpler and straighter solution is not yet developed.