PY-24220 Properly select minimum number of blanks after top-level imports
Namely, select the maximum number of empty lines between the settings
for the minimum number of blank lines surrounding top-level classes and
functions and the one for top-level imports.

We could give precedence to the preference for top-level imports, but it
would break existing formatting and lead to numerous pycodestyle.py
complaints because it defaults to only one blank line, and, according
to the PEP, there should be two in case of functions and classes
immediately following the import block.