diff: remove duplicated "Show Diff" actions from FindAction
These actions reuse shortcuts and presentation of each other.
This leads to confusion, when "Find Action" has 5 different
actions with same presentation.

This is a serious issue, as some of them can be enabled
in the same context, doing different things.
And if the shortcut invoked, the "wrong" action might happen.
For example: "Shelve" toolwindow had two actions available
(one is Shelve-specific, and another just finds List<Change>
in DataContext).