PY-23623 Fixed: False positive: __getitem__ not being defined for list after isinstance check
While calculating severity, we infer type for qualifier in `v[0]` subscription expression.
While inferring type for qualifier, we infer type for `test_if_list` function.
While inferring type for `test_if_list` function, we infer `Any` type for `v` in `v[0]` to prevent recursion and resolve `[]` to nothing.
Unresolved `[]` is cached so when we ask for `[]` on `v[0]`, we receive nothing but at this moment we know type of `v` (`List[Any]`).