Improve Update Project notification: oneline & with filter info
Mention just the whole number of affected files, no matter whether they were added, modified or deleted. This information is not critical in this place and, if needed, can be achieved from the update project tree.
This makes the notification oneline: IDEA-172375

If there is a scope filter, calculate the number of files under this filter, to let user know if he needs to view updated files at all: IDEA-172015.