[vcs-log] fix older index files not being removed
When a persistent structure is created for index, we provide it with a file that looks like "logId.kind.version". If this file does not exist, we assume that index version was incremented and attempt to delete files that start with "logId.kind". But this approach does not work with MapReduceIndex: it adds ".storage" suffix to the file given to it. In order to avoid deleting index files all the time, a brilliant solution was introduced in 91c7d97f0901bc97733a81c595c51ea51e0a0009: do not remove older versions at all. While the correct solution is to remove files that start with prefix corresponding to the older versions, which is done here.