PY-16683 Don't consider all decorator calls as having one implicit argument
It's not true, in particular, decorators of form @func(1, 2) don't
receive function object as an additional argument preceding those that
explicitly mentioned. Because of that we can't match arguments
of a decorator and parameters of the corresponding function definition,
e.g. while performing Change Signature.

There also seem to be no tests covering this pecular logic.