root index: compute dependent unloaded modules
We need to show a warning about possible incomplete results if there is an unloaded module which depends on a module from the project and e.g. 'Find Usages' is invoked on an element from such module. In order to implement that we need to quickly determine if there exists an unloading module which depends on a given loaded module.

(cherry picked from commit 9bc6d2f, IDEA-CR-21717)