Ensure Studio PSQ profile is a subset of Studio IDE local.
This is a one-off sync to the profiles. There is a hook
being added by ag/2394341 to ensure profiles are sync'd
on every change to them so manual sync wouldn't be required.

Note: This includes attributes reordering as this is how python’s
etree dumps xml. This impacts diff, but the actual difference in 
inspections is about 30 of them we needed to sync so far.

Test: N/A
Change-Id: Ie14d29ec5f3b2e5687dae2f3bf067e8853147ecb