[vcs-log] full indexing indexes everything
If a storage for project is fresh, indexing is not started until full log is read. Full indexing request indexes all commits, independently of the commits set passed to it.
Commits set is only used if full indexing is cancelled: a one-by-one request is scheduled for these commits in this case.