[vcs-log] avoid extra unnecessary validate requests
stateChanged method from ToolWindowManagerListener is called in a variety of the situations, most of which are not interesting from the log point of view (like focus moved in or out of the tool window, or other toolwindow closed). The only thing that is important is closing or opening of the vcs tool window. But before this commit, all even irrelevant changes triggered a validate request to the log. Previously extra requests that were sent by this method were not a problem, but now every new request cancels previous ones, so for example, a validate request sent by a focus moving out of the tool window can cancel filtering in the log, which is undesirable. So a check is added to detect vcs tool window opening or closing and nothing is done when a change does not concern vcs log.