Do not update editor history during dispose.
Bug: 37138939

The Intellij file TextEditorComponent.dispose() would update the
history, which in turn would update the last used editor to the
current editor. The problem is that in this case the text editor is
already removed from the list of tabs in a
EditorWithProviderComposite which is the component that is handling
several editors for the same file.

This is causing the XML text editor in Android Studio to store the
first editor (the design editor) instead of the current (the text
editor) as the most recent editor for the file.

The fix is a change in the text editor. The line removed was added
in 5/31/2013 by Zhdanov as a fix for this bug: IDEA-108057
"Caret position is messed up on navigation if editor tabs are off"
And the fix was to force to store editor settings on closing a file
at 'no tabs' mode.

To make sure this problem is still fixed, the history entry must be
updated before we dispose the EditorWithProviderComposite in
EditorWindow.clear().

I have logged the following bug: IDEA-175089 with Intellij.
And suggested the fix from this CL.

Test: Added a UI regression test.

Change-Id: Ic84c5a2a5be5e07809b69ca895cf36a24caa5eda