Move "annotation-internal" dependency from "util" to "core-api"
Although "util" module already contains two existing internal annotations, we cannot put the new, Java 8 annotations there (because of the Java version mismatch).

Also, we cannot add "annotation-internal" dependency to the "util" module (because neither core module could depend on non-core module, nor a module can depend on another module with a higher Java version).

Let's then add a dependency on the Java 8 annotations not to "util", but to "core-api" module (which also contains a related PathUtil), but still place the classes not in "openapi.jar", but in "util.jar" (because these annotations are "internal").