Revamp studio presubmit Java profile basing on errorprone mapping.
This CL wipes out the previous Java presubmit profile settings and
bases the new ones on the IntelliJ counterparts of errorprone errors.
See http://errorprone.info/bugpatterns for the full list of errorprone
issues, and the list of inspections below which were added on top of
that.

By this revamp, we ensure reusing google3 aggregated data as to which
inspections are most useful, while still retaining smooth IDE
integration for most of them, and reusing the high-precision useful
IntelliJ inspections which were not otherwise present in errorprone.

The rest of the inspections, including Kotlin, remain as per JetBrains
defaults.

This is the list of the inspections enabled on top of what errorprone
provides:

+ [ERROR] Math.random() cast to ‘int’
+ [ERROR] replaceAll(“.”, …)
+ [ERROR] Integer division in floating point context
+ [WARNING] Floating point comparison with == or !=
+ [WARNING] Assignment to method parameter
+ [WARNING] Assignment to lambda parameter
+ [WARNING] Nullability annotation problems
+ [WARNING] ConstantConditions

These are enabled in IntelliJ but have experimental status in
errorprone:
+ [ERROR] Divide by zero
+ [ERROR] Numbers comparison using == instead of equals()

The following have indirect matches with IntelliJ:

+ indexOf one-char string is replaceable by indexOf char (errorprone
checks the order of arguments though - so it's a different check, but
still it seemed reasonable to leave it as a warning - it's precise
enough and not disruptive, and I've seen some benchmarkes where it
can deliver slight performance improvement)
+ size() or length() comparison with zero is replaceable by isEmpty().
(errorprone one checks for size() >= 0, which is a different check
and would fall into the ConstantConditions one in IJ. But
still it has some intersection and having it in place seems reasonable,
not to mention real performance differences for collections which
don't store size as a field and its calculation could be O(N))

Those issues which didn't have IJ counterparts remain obviously not
enabled.

Another exception is MissingCasesInEnumSwitch - this CL demotes it to
a warning, like we do now in errorprone since this issue isn't auto-
fixable across the codebase and has many occurences to go through
manually. We don't want to fail PSQ because of it therefore, but want
people to look into it.

This CL also does the following:
1) Resets idea_default_no_spellchecker profile to the previous default,
and amends it then as per the pre-upload sync-inspection-profiles hook.
2) Sets idea_default_no_spellchecker as the IDE default. It appears
the hook operations were done on the no_spellchecker version, whereas
idea_default was the IDE profile - and this was likely unintentional.

Bug: b/63071706
Test: N/A
Change-Id: Iff79036a19ce690537aa4f251d1cb929227ef9f0