[ui] highlight usages color should get brighter for darcula, not darker
When cursor is on a variable, its usages are highlighted and gutter marks are shown. But after "Highlight usages in file" action is invoked, those marks become darker. This is fine in light ui themes, but hard to use in dark themes since dark colors are indistinguishable from dark background. So in Darcula color should become brighter, not darker.
This fixes IDEA-168493.