'bootstrap' module migrated to Java 8
It was left compatible with Java 6 to show a meaningful error message if user tries to start the IDE under Java 6/7. Given that we migrated to Java 8 a long time ago and bundle Java 8 JRE with all our IDEs this seems not actual anymore. 'bootstrap' module heavily depends on 'util' module so we need to perform this migration to migrate 'util' module to Java 8 as well.