Don't empty borders in the Git push options panel
If the border is empty, the focus border is empty as well.

On the other hand, insets() already take the border into account. There are a couple of extra pixels on Mac though, but seems to be a minor issue.

(cherry picked from commit a124e615461ef8fb81fdc3fcc458f26cd48bb21b)