Revert latest changes in PyDocStringTypeProvider and update its `getGenericType` method.
Consider class that is described in skeletons using legacy type hints syntax.
When somebody calls `getGenericType` on PyDocStringTypeProvider with this class as a first argument,
previous version of provider will take `__init__` from sources and returns NoneType.
But it should ask for return type of this `__init__` so PyUserSkeletonsTypeProvider will be called.
When PyUserSkeletonsTypeProvider will be called, it will load skeleton and then ask for skeleton return type.
After that `getReturnType` on PyDocStringTypeProvider will be called.

So it's important to not ignore `__init__` in PyDocStringTypeProvider and return specified value as is.

It's also important to filter return type in `getGenericType` method and return only appropriate one.

DjangoFormsTest.testFormFields and DjangoColumnsTest.testFields fixed.

(cherry picked from commit 72c8368)