[gui-test] change project output path if needed; cleanup; fixes
Build script output could interfere with project output path if we starting GUI tests with GuiTestLocalLauncher. It builds a classpath from main IDE and testGuiFramework modules output paths, so we should be careful with it and change a project output if needed (see GuiTestLocalLauncher#changeProjectOutputIfNeeded()). To detect a using of a build script output we are looking at URL for GuiTestLocalLauncher to load this class. If it is not same to path calculated from project model we are assuming that output path from build script is used and changing the project's output on it.

Groovy build scripts (CompilationContextImpl.groovy, BuildOptions.grovy) are reverting to its initial states.