Comment out problematic logging.
This causes occasional ConcurrentModificationException when starting the
native debugger.  It happens only if we need to resort to the code
injection strategy for starting the native debugger.

I've filed https://youtrack.jetbrains.com/issue/IDEA-175529 to get a
proper fix, but this will prevent the problem for now at very minor
cost.

Test: Existing
Change-Id: Id8f3807144d4f5017afbaff9bd329f51c54ac6a6