Use Kotlin plugin when running Studio locally
... either during unit/UI tests or when launched from Intellij.

There are two notable changes in PluginManagerCore, in order to
allow us to use -Dplugin.path to point to the Kotlin prebuilt:
- we have to disable the Maven plugin in all the above scenarios; we
do no ship it with Studio anyway, and the Kotlin plugin has a broken
optional dependency on it.
- we keep Kotlin in its separate classloader (when running locally,
typically the classpath already contains the necessary plugins and
the IDE will just instantiate one classloader for the entire thing).

Test: N/A
Change-Id: I97c0b54a8c3b24064e28d9c3e29463b8e3c5b2ad