[git] explicitly set diff.renameLimit when indexing
A big diff.renameLimit value set by user can slow down indexing tremendously. On the other hand, user may want to see renames in big commits when looking at them in the ui. Solution is to set the limit only when indexing. That is why parameter "fast" is introduced. It controls whether all renames are detected, or some of them are skipped in order to calculate details faster.

HgLogProvider ignores this parameter, since currently index for mercurial repositories is not built, but later is going to use it as well, since detecting moves/renames in hg is rather slow.

(cherry picked from commit 5dbf5bd)