Uses native temp file creation
This should fix the issue that prevents us from creating patches
in parallel. The old way when two patchers run at the same time
will likely return the same file for both to use.

Test: There are unit tests for the patcher. But they are all broken
and or failing before this change. Will follow up with Joe on this.
After my change no new tests are failing.

Change-Id: Id935386a210473c964b685f3902efab393936517