PY-24619: emulate qname for files with dashes to support them in tests
When user clicks on some element PyCharm gets its qname and generates
test target based on it. But user may use file with dashes.
There is no qName in this case, but all test runners support such targets.
So we emulate it in "getEmulatedQNameParts"

(cherry picked from commit 946dd6d)