PY-21002: Do not access remote servers from AWT thread PY-21002: Do not list interpreter packages under write action
We need to access remote server to check if it has Django.
We now pretend it has and check only when remote project creation
process starts because it runs under progress anyway.

PyUtil.runWithProgress executes its task on pool in all cases but when called under write action. In this case AWT thread is used.
There is nothing wrong with calling refreshAndGetPackagesModally on AWT unless we use write action because we need to use modal dialog anyway and that is ecactly what this method does.

Conclusion : we shall not use refreshAndGetPackagesModally under write action and all other usages are absolutely legal