avoid attach listener/write action race which could lead to unresponsiveness
When we try to attach app listener in the unfortunate moment between application firing beforeWriteActionStart() and actual write action start, all attached listeners won't be fired during this write action.
To avoid that, we check that if after attaching listener the app is still trying to start write action, then we declare the moment
unfortunate, detach listener and try again.