PY-25106 Fix ipnb file-level metadata parsing, mainly codemirror and nbformat_minor options
Before this change we parsed metadata as a map, but it failed to handle ipython "version" argument in "codemirror_mode" options, because Gson parse all numbers ad double by default. As it's not easily customisable, I decided to register type adapter for deserialization. To address the problem with version I manually parse "language_info" and "codemirror_mode" options respectively and get "version" option as Int.