[git] collect "incomplete" merge commits and process them together
Due to an issue with -m parameter, in order to index merge commits that are same as one of their parents, their and their parents tree hashes are requested with additional "git log" calls.
In order to minimise number of required calls, its better to process such commits together. Ideally we'd like to do it at the very end. But since for each such commit a record is kept until it is processed,
 processing also is called when too much data (in terms of status lines) is accumulated.

IDEA-176504

(cherry picked from commit eed01c2)