Expire Update notification, when info tab is closed
We could recreate the info tab, when user clicks "View" again
(as it was attempted in e017f4a), but it would require storing
UpdatedFiles information in memory during indefinite amount of time
(without any additional actions: until project close).

On the other hand, if user explicitly closes the tab, probably he doesn't
need it anymore, and it is safe to drop the content.

This fixes IDEA-176051
Issues found in IDEA-CR-23204