[vcs-log] ability to turn indexing off for the project with an override switch in registry
Some projects have repositories that are too enormous to index (main problem is having commits with hundred thousands changes). So the ability to switch indexing off for such project is introduced.
To do so a line should be added to .idea/vcs.xml:

<component name="Vcs.Log.Settings" index="false" />

This switch can be overridden by a registry property vcs.log.index.force if a user is willing to try his luck.
Since there is only one candidate project to switch off indexes, no ui is introduced right now.