Fix Git interactive rebase and reword dialogs' issues with lines starting with a hash sign
Override core.commentChar with our special "\u0001". Given that it is highly improbably that a line in a commit message starts with this symbol, this allows to treat it as a comment.

This allows to fix 3 issues:
* IDEA-176455 and IDEA-175443: Don't loose lines starting with a hash (the hash is not a comment char anymore).
* IDEA-140568: Ignore lines, passed to the Git interactive rebase dialog, which start with a redefined commentChar (we redefine it again to \u0001 and ignore lines starting with that specific symbol).
* Don't show Git helping text in reword/squash dialogs (again by ignoring lines starting with \u0001).

All issues remain in the very rare case, when a meaningful line starts with \u0001, but it was decided to ignore this issue until reported. If we need to fix it further, the "cleanup=whitespace" configuration option can be used to keep the lines starting with \u0001 (Git won't stripe them out on commit), but a more intelligent parsing will be needed to show these lines in the reword/squash dialog.